Add support for Module Scopes
The scope of a modules specifies its lifecycle.

By default, a module is instance scoped, i.e. it will be instantiated each time we request an instance of the agent from the Registry.

You can also specify a module to be agent scoped. In that case it will be instantiated immediately upon agent registration, and all instances of the agent will use the same instance of this module.

By registering a module at the server level, you can also have a server scoped module. Those can not be defined at the agent level, but they can be used there. For example a module could depend on Xaibo itself and that way interact with other known agents.